refactor(brain-core): profile self-update via apply_signal
Move BrainSignal to brain-core and add apply_signal() methods to
BalancedRecallState and SectionPosteriorState. Each profile state type
now owns its update logic — the handler dispatches a single call
rather than hardcoding per-type fold instantiation.

- brain_signal.rs: BrainSignal enum + entity_signal/is_recall_positive
- profile.rs: BalancedRecallState::apply_signal (from BalancedRecallFold)
- section_state.rs: SectionPosteriorState::apply_signal (from SectionPosteriorFold)
- fold.rs: thin adapter (interpret → apply_signal) for Fold<Event,S> compat
- handlers.rs: one apply_signal call replaces fold instantiation
- persist.rs: replay uses apply_signal directly, no fold imports
- event.rs: re-exports BrainSignal from brain-core, keeps interpret()

Net -221 lines. Profile types are now self-contained — adding a new
profile class means implementing apply_signal on its state type, no
handler edits needed.
